AR-325: measure replication delay

+++ src/main/java/gov/ca/cwds/neutron/flight/FlightLog.java
@@ -860,11 +860,11 @@ public class FlightLog implements ApiMarker, AtomRocketControl {
       attribs.putIfAbsent("run_start_time", Instant.ofEpochMilli(startTime).getEpochSecond());
       attribs.putIfAbsent("run_end_time", Instant.ofEpochMilli(endTime).getEpochSecond());
 
-      final long runSeconds = (endTime - startTime) / 1000;
+      final float runSeconds = (endTime - startTime) / 1000;
       LOGGER.debug("Neutron: this run seconds: {}", runSeconds);
       attribs.putIfAbsent("run_seconds", runSeconds);
 
-      final long runMillis = endTime - startTime;
+      final float runMillis = endTime - startTime;
       attribs.putIfAbsent("run_millis", runMillis);
 
       // AR-325: replication metrics.
@@ -873,7 +873,7 @@ public class FlightLog implements ApiMarker, AtomRocketControl {
       if (lastEndTime != 0) {
         attribs.putIfAbsent("last_run_end_time",
             Instant.ofEpochMilli(lastEndTime).getEpochSecond());
-        final long runTotalMillis = lastEndTime - startTime;
+        final float runTotalMillis = lastEndTime - startTime;
         final float runTotalSeconds = runTotalMillis / 1000;
 
         LOGGER.debug("since last run: millis: {}, seconds: {}", runTotalMillis, runTotalSeconds);
